Added additional filtering
Contour / ellipse test now has an addition step that requires the contour to actually be an ellipse vs a partial one. This further reduces the false alarms and it is nearly perfect using Brad Miller's field pictures. All of them; not just the few in GitHub.

Still need to test on a live camera with different lighting conditions.